Use new API from confluentinc/kafka-connect-storage-common#77
The API change itself is trivial to adopt. Get the original values, add
the defaults, and that's done.

However, many tests either relied on implementation details of the
partitioners, or ignored the mechanism used in the very same tests to
pass settings in a reliable way and, instead, modified the configuration
right before it was passed to the partitioner, which made it unavailable
elsewhere as code migrated to relying on HdfsSinkConnectorConfig instead
of the no-longer present parsed config map.